requirements.txt
numpy==1.21.6
pytesseract==0.3.10
PyAutoGUI==0.9.53
pyperclip==1.8.2
keyboard==0.13.5
Pillow==9.2.0
python-dotenv==0.21.0
opencv-python==4.6.0.66
mss==6.1.0